+++ src/Events/JWTLogin.php
@@ -11,31 +11,6 @@
 
 namespace Tymon\JWTAuth\Events;
 
-class JWTLogin
+class JWTLogin extends AbstractJWTEvent
 {
-    /**
-     * The authenticated user.
-     *
-     * @var \Illuminate\Contracts\Auth\Authenticatable
-     */
-    public $user;
-
-    /**
-     * @var \Tymon\JWTAuth\Token
-     */
-    public $token;
-
-    /**
-     * Create a new event instance.
-     *
-     * @param  \Illuminate\Contracts\Auth\Authenticatable  $user
-     * @param  \Tymon\JWTAuth\Token|string  $token
-     *
-     * @return void
-     */
-    public function __construct($user, $token)
-    {
-        $this->token = $token;
-        $this->user = $user;
-    }
 }
